requirements.txt
ultralytics
supervision
opencv-python
numPy
matplotlib
pandas
roboflow
python-dotenv